feat(as-4185): addressed pr comments
