+++ packages/lpa-questionnaire-web-app/__tests__/unit/full-appeal/mockAppeal.js
@@ -1,5 +1,3 @@
-
-
 const {
   constants: { APPEAL_ID, APPLICATION_DECISION },
 } = require('@pins/business-rules');
